+++ src/Tests/Dfe.Complete.Api.Tests.Integration/Controllers/ProjectsController/TaskDataControllerTests.cs
@@ -45,7 +45,7 @@ namespace Dfe.Complete.Api.Tests.Integration.Controllers.ProjectsController
         public async Task UpdateHandoverWithDeliveryOfficerTaskDataByProjectIdAsync_ShouldUpdate_ConversionTaskData(
             CustomWebApplicationDbContextFactory<Program> factory,
             ITasksDataClient tasksDataClient,
-            UpdateHandoverWithDeliveryOfficerCommand command,
+            Complete.Client.Contracts.UpdateHandoverWithDeliveryOfficerTaskCommand command,
             IFixture fixture)
         {
             // Arrange
@@ -78,7 +78,7 @@ namespace Dfe.Complete.Api.Tests.Integration.Controllers.ProjectsController
         public async Task UpdateHandoverWithDeliveryOfficerTaskDataByProjectIdAsync_ShouldUpdate_TransferTaskData(
             CustomWebApplicationDbContextFactory<Program> factory,
             ITasksDataClient tasksDataClient,
-            UpdateHandoverWithDeliveryOfficerCommand command,
+            UpdateHandoverWithDeliveryOfficerTaskCommand command,
             IFixture fixture)
         {
             // Arrange
@@ -110,7 +110,7 @@ namespace Dfe.Complete.Api.Tests.Integration.Controllers.ProjectsController
         } 
         [Theory]
         [CustomAutoData(typeof(CustomWebApplicationDbContextFactoryCustomization))]
-        public async Task GetConversionTasksDataByIdAsync_ShouldReturn_A_TaskDaa(
+        public async Task GetConversionTasksDataByIdAsync_ShouldReturn_A_TaskData(
         CustomWebApplicationDbContextFactory<Program> factory,
         ITasksDataClient tasksDatClient,
         IFixture fixture)
@@ -207,6 +207,40 @@ namespace Dfe.Complete.Api.Tests.Integration.Controllers.ProjectsController
             Assert.True(existingTaskData.SupplementalFundingAgreementCleared);
             Assert.True(existingTaskData.SupplementalFundingAgreementReceived);
             Assert.False(existingTaskData.SupplementalFundingAgreementSaved); 
+        }
+        [Theory]
+        [CustomAutoData(typeof(CustomWebApplicationDbContextFactoryCustomization))]
+        public async Task UpdateArticleOfAssociationTaskAsync_ShouldUpdate_ConversionTaskData(
+            CustomWebApplicationDbContextFactory<Program> factory,
+            ITasksDataClient tasksDataClient,
+            UpdateArticleOfAssociationTaskCommand command,
+            IFixture fixture)
+        {
+            // Arrange
+            factory.TestClaims = [new Claim(ClaimTypes.Role, ApiRoles.ReadRole), new Claim(ClaimTypes.Role, ApiRoles.UpdateRole), new Claim(ClaimTypes.Role, ApiRoles.WriteRole)];
+
+            var dbContext = factory.GetDbContext<CompleteContext>();
+
+            var taskData = fixture.Create<ConversionTasksData>();
+            dbContext.ConversionTasksData.Add(taskData);
+
+            await dbContext.SaveChangesAsync();
+            command.TaskDataId = new TaskDataId { Value = taskData.Id.Value };
+            command.ProjectType = ProjectType.Conversion;
+            command.NotApplicable = true;
+
+            // Act
+            await tasksDataClient.UpdateArticleOfAssociationTaskAsync(command, default);
+
+            // Assert
+            dbContext.ChangeTracker.Clear();
+            var existingTaskData = await dbContext.ConversionTasksData.SingleOrDefaultAsync(x => x.Id == taskData.Id);
+            Assert.NotNull(existingTaskData);
+            Assert.True(existingTaskData.ArticlesOfAssociationNotApplicable);
+            Assert.Null(existingTaskData.ArticlesOfAssociationCleared);
+            Assert.Null(existingTaskData.ArticlesOfAssociationReceived);
+            Assert.Null(existingTaskData.ArticlesOfAssociationSaved);
+            Assert.Null(existingTaskData.ArticlesOfAssociationSent);
         }
     }
 }